// src/app/shared/models/survey.model.ts
export class SurveyModel {
    street: any;
    pincode: any;
    area: any;
    houseno: any;
    apartment: any;
    email: any;
    name: any;
    country: any;
    city: any;
    lat: any;
    lang: any;
    doyoucompost: any;
    doyoucompostAns: any;
    apartmentcommunitylanehouseholdunit: any;
    ifathomeans: any;
    howlongcomposting: any;
    whatcomposterdoyouuse: any;
    doyoudowithyourcompost: any;
    doyouhaveexcesscompost: any;
    willingsignupswachagraha: any;
    doyouexclusivelycompostleavesflowers: any;
    leavesflowerscompostertype: any;
    stayconnectedwhatsapp: any;
    contactno: any;
    wardnumber: any;
    wardname: any;
    constituencyname: any;
    constituencynumber: any;
    // tslint:disable-next-line:variable-name
    formatted_address: any;
    timestamp: any;
    state: any;
    source: any;
    customertype: any;
}

export class ExportExcel extends SurveyModel {
    Key: any;
}